separate mapbox overriding style
